requirements.txt
numpy==1.18.1
pandas==1.0.3
praat-parselmouth==0.3.3
scipy==1.4.1
pydub==0.24.1